test: fix the test cases for throttleTime
Seems like the test case was written for the sampleTime strategy. The elapsed time between `start` end `end` was near 0ms because ThrottleTime emits a value then ignores other values for a duration (when leading is true, trailing is false). Therefore the expected elapsed time for the test nammed `should create a Observable of scroll observable with default 300 throttle` could not be 300ms.